pyproject.toml
[project]
name = "mcp-dcma"
version = "0.1.0"
description = ""
authors = [
    {name = "KangHyeon3315",email = "son2786w@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
